requirements.txt
certifi==2017.7.27.1
chardet==3.0.4
click==6.7
Flask==0.11.1
idna==2.6
itsdangerous==0.24
Jinja2==2.9.6
MarkupSafe==1.0
PyYAML==3.12
requests==2.18.4
urllib3==1.22
virtualenv==15.1.0
Werkzeug==0.12.2
deap==1.0.2
